requirements.txt
gunicorn
Werkzeug
openai
flask
google-cloud-firestore
google-auth